api/requirements.txt
fastapi==0.95.2
uvicorn
python-multipart
gdown==4.6.4
numpy==1.23.3
onnxruntime==1.14.1
Pillow==9.2.0
transformers==4.28.1